添加重置密码功能

+++ backend/service/code_service.py
@@ -9,17 +9,22 @@ class CaptchaService:
         self.dao = CaptchaDAO()
 
     def validate(self, email: str, code: str) -> bool:
-        # 查询验证码
+        self.dao.session.expire_all()  # 强制刷新
         res_code = self.dao.query_captcha_code(email)
-
         now = datetime.now().timestamp()
-        self.dao.delete_captcha(email)
-        # 对比数据库时间戳是否超时，若超时则返回
-        if (
-            res_code is not None
-            and code == res_code.verify_code
-            and now - res_code.verify_code < 60 * 10
-        ):
-            return True
+        # print(f"当前时间：{now}")
+        if res_code:
+            pass
+            # print(f"验证码记录：{res_code.verify_code}, 生成时间：{res_code.verify_time}, 差值：{now - res_code.verify_time}")
         else:
-            return False
+            print("没有查询到验证码记录")
+        if res_code is None:
+            return True
+        if code != res_code.verify_code:
+            return True
+        if now - res_code.verify_time >= 60 * 10:
+            return True
+        # 验证通过后删除验证码记录
+        self.dao.delete_captcha(email)
+        return False
+
